Fix issue with "about" page introduced with theme upgrade
The about/list.html in the theme would take precedence over our
section/about.html, so rename our to override the theme one.

Do the same change for "partner" for consistency, although
there was no conflict for that one, since only "about" is part
of the original theme.

The issue probably appeared with commit 58914957 in the
hugo-creative-portfolio-theme repository, which also bumped the
hugo version required to 0.20.